+++ CHANGES.md
@@ -3,6 +3,12 @@
 
 This is a high-level summary of the most important changes. For a full list of changes, see the [git commit log](https://github.com/grindsa/dkb-robo/commits) and pick the appropriate release branch.
 
+## Unreleased
+
+**Bugfixes**:
+
+- Forward new `headless` option from CLI and DKBRobo into captcha token solving
+
 ## Changes in 0.31.1
 
 - [#89](https://github.com/grindsa/dkb-robo/pull/89) - Postbox scan: HTTPError: 401 Client Error: Unauthorized